Move discussions API routes to separate directory
